fix build errors one windows and correct var types (FSBUILD-225)
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@16112 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/switch_time.c
@@ -69,7 +69,7 @@ static int MATRIX = 1;
 
 static int STEP_MS = 10;
 static int STEP_MIC = 10000;
-static int TICK_PER_SEC = 100;
+static uint32_t TICK_PER_SEC = 100;
 
 static int MS_PER_TICK = 10;
 
@@ -185,7 +185,7 @@ static void calibrate_clock(void)
 	for (x = 0; x < 500; x++) {
 		avg = average_time(val, 100);
 		
-		if (abs(want - avg) <= 2) {
+		if (abs((int)(want - avg)) <= 2) {
 			if (++good > 10) {
 				break;
 			}
@@ -198,7 +198,7 @@ static void calibrate_clock(void)
 		}
 	}
 
-	OFFSET = want - val;
+	OFFSET = (int)(want - val);
 }
